# Don't read config in a tight loop