When checking if we need to output to the screen, we had accidentally been checking the config at each loop step as it was the first step of `eval_block`. Instead, check the config closer to where the output will actually happen so we only pay for the lookup when we're already going to pay for an i/o.