@fdncred 2021-12-28
before this pr on my Mac
```
〉benchmark { 1..100000 | each { $it + 10 } | length }
5sec 331ms 467us 750ns
```
after
```
〉benchmark { 1..100000 | each { $it + 10 } | length }
164ms 332us 125ns
```