@jntrnr 2021-12-28
@fdncred what does nushell do with `benchmark { echo 1..100000 | each { $it + 10 } | length }` on your mac?