@fdncred 2021-12-28
@jntrnr 
```
benchmark { echo 1..100000 | each { $it + 10 } | length }
╭───┬──────────────────╮
│ # │ real time        │
├───┼──────────────────┤
│ 0 │ 72ms 565us 250ns │
╰───┴──────────────────╯
```